fix: refactor of topic component
