+++ src/pages/topics/ui/stage-tabs/index.tsx
@@ -1,4 +1,5 @@
-import { TopicGrid, StageOverview } from '../index';
+import { StageOverview } from '../stage-overview';
+import { TopicGrid } from '../topics-grid';
 
 import { STAGES, DEFAULT_STAGES_VALUE, Tabs, TabsList, TabsTrigger, TabsContent } from '@/shared';
 import type { Topic, TopicProgress } from '@/shared/api';
@@ -17,7 +18,7 @@ export const StageTabs = ({
   onSelectionChange,
 }: StageTabsProps) => {
   return (
-    <section className="container mx-auto max-w-7xl px-4 pt-6 gap-2">
+    <section className="container mx-auto max-w-7xl px-4 pt-6">
       <Tabs defaultValue={String(DEFAULT_STAGES_VALUE)} className="w-full mx-auto max-w-312">
         <TabsList variant="stages">
           {STAGES.map((stage) => (
@@ -27,7 +28,7 @@ export const StageTabs = ({
           ))}
         </TabsList>
         {STAGES.map((stage) => (
-          <TabsContent key={stage.id} value={String(stage.id)}>
+          <TabsContent key={stage.id} value={String(stage.id)} className="space-y-4">
             <StageOverview
               stageId={stage.id}
               totalTopics={groupedTopics[stage.id]?.length ?? 0}
